branch   interlude     lunos armor set passive skills
Basicly it's adding passive skills to armor sets (now it's hardcoded in xmls and actually is not passive skills). With that patch you can define armor sets and attach passive skill to it (when full armor set is uquiped then passive skill should be activated [ and be seen in skill list] ).
its beeing tested (needs dp update) and big thx to luno :)
http://forum.l2jserver.com/thread.php?threadid=22214


git-svn-id: http://svn.l2jfree.com/l2jfree@294 6ff00ba9-25cd-4355-987e-099c62bc2628