trunk 	gs 		sync to l2j 899
Quest/AI minor enhancements
Addition of onAdvEvent which works exactly the same as onEvent, except passes possibly null npc and player instances associated with the event.
For protection against hlapex, a quest state is still required from a player in bypasses, but not in timers.
onAdvEvent will call onEvent if not implemented explicitely by a quest.

Moving the timers to Quest so that they may be started even when the player has no quest state...or even when no player starts them.

All changes are backwards compatible with old quest scripts. 


git-svn-id: http://svn.l2jfree.com/l2jfree@1522 6ff00ba9-25cd-4355-987e-099c62bc2628